Move BlockDagKeyValueStorage to Casper project
