+++ casper/src/test/scala/coop/rchain/casper/helper/BlockDagStorageFixture.scala
@@ -4,11 +4,8 @@ import cats.effect.Concurrent
 import cats.syntax.all._
 import coop.rchain.blockstorage.blockStore
 import coop.rchain.blockstorage.blockStore.BlockStore
-import coop.rchain.blockstorage.dag.{
-  BlockDagKeyValueStorage,
-  BlockDagStorage,
-  IndexedBlockDagStorage
-}
+import coop.rchain.blockstorage.dag.{BlockDagStorage, IndexedBlockDagStorage}
+import coop.rchain.casper.dag.BlockDagKeyValueStorage
 import coop.rchain.casper.storage.RNodeKeyValueStoreManager
 import coop.rchain.casper.util.GenesisBuilder.GenesisContext
 import coop.rchain.casper.util.rholang.{Resources, RuntimeManager}
